chore: load Rails configuration defaults for combustion app
This was causing our system specs to fail as
`config.action_view.form_with_generates_ids` was not being set, which
caused issues with capybara linking labels with fields.

In order to fix it, I think the best approach is to simply load all the
config defaults of the current Rails version.