fix: avoid assuming `email` as the authentication key of the resource
Addresses #63.

This commits fixes an issue where our form helpers for creating
credentials and for authenticating with them weren't working in
applications that doesn't use `email`s for identifying their resource –
for example, they use `username` instead.

In order to fix this, this commit makes it so that it looks for the
authentication keys of the resource and returns the first value obtained
after evaluating each of them.