feat: generate webauthn credentials table with not null constraints for attributes that must be present (#70)
* feat: add not null constraints to webauthn credentials attributes

This commit makes it so that the generated migration for creating
webauthn credentials now has not null constraints in
`webauthn_credentials` attributes `external_id`, `name`, `public_key`,
`sign_count` and `authentication_factor`.

In order to do that, it uses the new not-null modifier (`!`) for active
record generators if using Rails 8 or it uses a newly added migration
template if using Rails < 8.

References:
- https://github.com/rails/rails/commit/2f92b1c94e2a080679ab0ad45abd0bdec24a5349

* refactor: keep it simple and always use migration template

* test: fix tests

* test: update testing app database schema

* docs: update `CHANGELOG.md`

* chore: fix typo